feat: move extension to Core(#301)
