+++ 14th-team5-iOS/Core/Sources/Extensions/Array+Ext.swift
@@ -1,0 +1,14 @@
+//
+//  Array+Ext.swift
+//  Core
+//
+//  Created by 마경미 on 20.01.24.
+//
+
+import Foundation
+
+extension Array {
+    public init(with element: Element) {
+        self = [element]
+    }
+}
